Extract the Gleam JSON component
- Move JSON storage, parsing, and provider ownership into geam-json.
- Consume stdlib internals through one narrow support contract.
- Preserve the root JSON facade and official compatibility behavior.
