+++ Cargo.toml
@@ -19,8 +19,8 @@ include = [
 ]
 
 [workspace]
-members = [".", "core", "builtins/stdlib"]
-default-members = [".", "core", "builtins/stdlib"]
+members = [".", "core", "builtins/stdlib", "builtins/json"]
+default-members = [".", "core", "builtins/stdlib", "builtins/json"]
 resolver = "3"
 
 [workspace.package]
@@ -41,6 +41,7 @@ clap = { version = "4", features = ["derive"] }
 ecow = "=0.2.6"
 flate2 = "1"
 geam-core = { path = "core", version = "=0.1.2" }
+geam-json = { path = "builtins/json", version = "=0.1.2" }
 geam-stdlib = { path = "builtins/stdlib", version = "=0.1.2" }
 gleam-core = { package = "geam-gleam-core", version = "=1.18.1-geam.2" }
 gleam-compiler-core = { package = "geam-gleam-core", version = "=1.18.1-geam.2" }
@@ -76,6 +77,7 @@ clap.workspace = true
 ecow.workspace = true
 flate2.workspace = true
 geam-core.workspace = true
+geam-json.workspace = true
 geam-stdlib.workspace = true
 gleam-core.workspace = true
 half.workspace = true
@@ -83,12 +85,10 @@ hex.workspace = true
 hexpm.workspace = true
 im.workspace = true
 jiff.workspace = true
-jiter.workspace = true
 miette.workspace = true
 num-bigint.workspace = true
 num-traits.workspace = true
 rand.workspace = true
-ryu.workspace = true
 semver.workspace = true
 serde.workspace = true
 serde_json.workspace = true
